pyproject.toml
[project]
name = "py-framework"
version = "0.1.0"
description = "Secure FastHTML and DuckDB web framework with authentication"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-fasthtml>=0.6.9",
    "duckdb>=1.1.0",
    "passlib[bcrypt]>=1.7.4,<2.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.9",
    "httpx>=0.27.0",
    "pydantic>=2.8.0",
    "pydantic-settings>=2.4.0",
    "python-dotenv>=1.0.0",
    "email-validator>=2.0.0",
    "jinja2>=3.1.4",
    "itsdangerous>=2.2.0",
    "cryptography>=42.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
    "ruff>=0.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/framework"]

[tool.black]
line-length = 88
target-version = ['py313']

[tool.ruff]
target-version = "py313"
line-length = 88

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
    "ruff>=0.5.0",
]
